Switch XorNames from ref-array to JS Array (#64)
* feat/Immutable XorName is now represented as an Array

Previously we were leaking out ref-array objects here, but that caused
troubles when they were used at other parts of the API. Namely if they
were serialised without being converted. In order to make things easier
we are now always copying them into a JS Array before returning them in
the API call and are copying back if they are given as such.

* fix/NFS convert incoming array back to ref-array

Internally we are using ref-array structs when dealing with data_map_names
and since the latest change this has been a JS array instead. This checks
for that when the (nfs) File Object is created and converts it into the
proper ref-array type again.